library_management_system
The Library Management System is a console-based Java application designed to help manage books in a library efficiently. It allows librarians or users to perform basic operations like adding new books, viewing all books, borrowing books, and returning borrowed books.

The system uses Object-Oriented Programming (OOP) concepts such as classes, objects, encapsulation, and lists to organize and manage the data. It is implemented without a database — all data is stored temporarily in memory (using an ArrayList).

This project is simple and ideal for beginners learning Java programming and object-oriented design.